Admin, User 로직 통합
