--- src/main/java/com/automart/admin/dto/AuthResponseDto.java
@@ -1,17 +1,0 @@
-package com.automart.admin.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class AuthResponseDto {
-
-    private String accessToken;
-    private String tokenType = "Bearer";
-
-    public AuthResponseDto(String accessToken) {
-        this.accessToken = accessToken;
-    }
-
-}
