+++ src/main/java/com/automart/user/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.automart.user.controller;
 
+import com.automart.advice.exception.InvalidTokenException;
 import com.automart.advice.exception.NotFoundUserException;
 import com.automart.advice.exception.SMSException;
 import com.automart.security.UserPrincipal;
@@ -16,12 +17,21 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
+import java.util.Date;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 @Api(tags = {"2. User Management"})
 @RestController
@@ -36,6 +46,15 @@ public class UserController {
     private UserRepository userRepository;
 
     @Autowired
+    private JwtTokenProvider jwtTokenProvider;
+
+    @Autowired
+    private AuthenticationManager authenticationManager;
+
+    @Autowired
+    private PasswordEncoder passwordEncoder;
+
+    @Autowired
     RedisTemplate<String, Object> redisTemplate;
 
 
@@ -44,6 +63,7 @@ public class UserController {
             @ApiResponse(code = 200, message = "일치하는 회원이 존재합니다."),
             @ApiResponse(code = 403, message = "일치하는 회원이 존재하지 않습니다.")
     })
+    @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/find/email")
     public ResponseEntity<String> findEmail(@ApiParam("이름, 휴대폰번호 정보") @Valid @RequestBody FindRequestDto requestDto) throws NotFoundUserException {
         User user = userService.findUserByNameAndTel(requestDto.getName(), requestDto.getPhone());
@@ -57,6 +77,7 @@ public class UserController {
             @ApiResponse(code = 200, message = "전송된 인증번호 반환"),
             @ApiResponse(code = 500, message = "메세지 전송 실패"),
     })
+    @PreAuthorize("hasRole('USER')")
     @PostMapping("/find/valid/phone")
     public ResponseEntity<Integer> validPhoneForFindPw(@RequestParam String phoneNo) throws SMSException {
         int validNum = userService.validatePhone(phoneNo);
@@ -70,6 +91,7 @@ public class UserController {
             @ApiResponse(code = 500, message = "메세지 전송 실패"),
             @ApiResponse(code = 403, message = "일치하는 회원이 존재하지 않습니다.")
     })
+    @PreAuthorize("hasRole('USER')")
     @PostMapping("/find/reissue/password")
     public ResponseEntity<Void> findPassword(@ApiParam("이름, 휴대폰번호 정보") @Valid @RequestBody FindRequestDto requestDto) throws NotFoundUserException, SMSException {
         User user = userService.findUserByNameAndTel(requestDto.getName(), requestDto.getPhone());
@@ -86,6 +108,7 @@ public class UserController {
             @ApiResponse(code = 200, message = "정상적으로 내정보가 불러와졌습니다."),
             @ApiResponse(code = 401, message = "토큰 만료", response = AuthResponseDto.class)
     })
+    @PreAuthorize("hasRole('USER')")
     @GetMapping("/me")
     public User getCurrentUser(@AuthenticationPrincipal UserPrincipal userPrincipal) {
         return userRepository.findByNo(userPrincipal.getNo())
@@ -97,6 +120,7 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 200, message = "정상적으로 해당 회원정보가 조회되었습니다.")
     })
+    @PreAuthorize("hasRole('ADMIN')")
     @GetMapping("/byEmail")
     public ResponseEntity<UserResponseDto> showUser(@ApiIgnore @RequestHeader("Authorization") String token,
                                                     @RequestParam String email) {
@@ -108,6 +132,7 @@ public class UserController {
     @ApiResponses({
             @ApiResponse(code = 200, message = "회원목록을 조회하는데 성공했습니다.")
     })
+    @PreAuthorize("hasRole('ADMIN')")
     @GetMapping(value = "/list")
     public ResponseEntity<List<UserResponseDto>> showUsers(@ApiIgnore @RequestHeader("Authorization") String token) {
         List<UserResponseDto> userResponseDtos = userService.showUsers();
@@ -115,15 +140,62 @@ public class UserController {
     }
 
 
-    /*@ApiOperation(value = "비밀번호 변경 전 확인", notes = "회원이 비밀번호 변경 전 비밀번호가 일치하는지 확인한다.", authorizations = { @Authorization(value = "jwtToken")})
+    @ApiOperation(value = "(개발용) 회원 탈퇴", notes = "특정 회원을 삭제한다", authorizations = { @Authorization(value = "jwtToken")})
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "회원이 정상적으로 탈퇴되었습니다.")
+    })
+    @PreAuthorize("hasRole('USER')")
+    @DeleteMapping(value = "/delete")
+    public ResponseEntity<String> withdraw(@ApiIgnore @RequestHeader("Authorization") String token) {
+        return ResponseEntity.status(HttpStatus.OK).body(userService.withdraw(token, JwtTokenProvider.TokenType.ACCESS_TOKEN));
+    }
+
+
+    @ApiOperation(value = "로그아웃", notes = "로그인된 계정을 로그아웃한다.", authorizations = { @Authorization(value = "jwtToken")})
+    @ApiResponse(code = 200, message = "로그아웃 되었습니다.")
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    @PostMapping(value = "/logout")
+    public ResponseEntity<Void> logout(@AuthenticationPrincipal UserPrincipal userPrincipal,
+                                       HttpServletRequest request) {
+        String accessToken = jwtTokenProvider.extractToken(request);
+        log.info("accessToken : "+accessToken);
+        String principal = null;
+
+        //* access token을 통해 userEmail or adminId를 찾아 redis에 저장된 refresh token을 삭제한다.*//
+        try {
+            principal = userPrincipal.getPrincipal();
+        } catch (InvalidTokenException e) {
+            log.error("userEmail or adminId가 유효한 토큰에 존재하지 않음.");
+        }
+
+        try {
+            if (redisTemplate.opsForValue().get(principal) != null) {
+                redisTemplate.delete(principal);
+            }
+        } catch (IllegalArgumentException e) {
+            redisTemplate.delete(principal);
+        }
+
+        //* access token이 유효한 토큰인 경우 더 이상 사용하지 못하게 블랙리스트에 등록 *//
+        if (jwtTokenProvider.validateToken(accessToken)) {
+            Date expirationDate = jwtTokenProvider.getExpirationDate(accessToken, JwtTokenProvider.TokenType.ACCESS_TOKEN);
+            redisTemplate.opsForValue().set(
+                    accessToken, true,
+                    expirationDate.getTime() - System.currentTimeMillis(), TimeUnit.MILLISECONDS); // 토큰의 유효기간이 지나면 자동 삭제
+            log.info("redis value : " + redisTemplate.opsForValue().get(accessToken));
+        }
+
+        return ResponseEntity.status(HttpStatus.OK).build();
+    }
+
+    @ApiOperation(value = "비밀번호 변경 전 확인", notes = "회원이 비밀번호 변경 전 비밀번호가 일치하는지 확인한다.", authorizations = { @Authorization(value = "jwtToken")})
     @ApiImplicitParam(name = "userPassword", value = "해당 회원의 현재 비밀번호", required = true, dataType = "string", defaultValue = "oldpassword")
     @ApiResponses({
             @ApiResponse(code = 200, message = "토큰에 해당하는 유저가 존재합니다.\n(비밀번호가 일치하면 true, 일치하지 않으면 false를 반환)"),
             @ApiResponse(code = 401, message = "토큰 만료", response = AuthResponseDto.class)
     })
-
     @PreAuthorize("hasRole('USER')")
-    @PostMapping(value = "/varifyPassword")
+    @PostMapping(value = "/valid/password")
     public ResponseEntity<Boolean> verifyPassword(@AuthenticationPrincipal UserPrincipal userPrincipal,
                                                   @RequestParam String userPassword) {
         User user = getCurrentUser(userPrincipal);
@@ -132,7 +204,44 @@ public class UserController {
             return ResponseEntity.status(HttpStatus.OK).body(false);
 
         return ResponseEntity.status(HttpStatus.OK).body(true);
-    }*/
+    }
+
+    @ApiOperation(value = "8 로그인 후 비밀번호 변경", notes = "로그인 후 임시비밀번호를 변경한다.", authorizations = {@Authorization(value = "jwtToken")})
+    @ApiImplicitParam(name = "newPassword", value = "해당 회원의 새 비밀번호", required = true, dataType = "string", defaultValue = "newpassword")
+    @ApiResponses({
+            @ApiResponse(code = 201, message = "정상적으로 비밀번호를 변경 후, 새로운 토큰을 발급한다."),
+            @ApiResponse(code = 401, message = "토큰 만료로 인해 비밀번호 변경 불가 -> 새로운 토큰 발급", response = AuthResponseDto.class),
+            @ApiResponse(code = 403, message = "일치하는 회원이 존재하지 않아 비밀번호 변경에 실패하였습니다.")
+    })
+    @PreAuthorize("hasRole('USER')")
+    @PostMapping(value = "/change/password")
+    public ResponseEntity<AuthResponseDto> changePassword(@ApiIgnore @RequestHeader("Authorization") String token,
+                                                          @AuthenticationPrincipal UserPrincipal userPrincipal,
+                                                          @RequestParam String newPassword) throws NotFoundUserException {
+        /* password 변경 */
+        User user = userService.changePassword(userPrincipal.getNo(), newPassword);
+        user.makeFalseTempPw();
+
+        /* access token이 유효한 토큰인 경우 더 이상 사용하지 못하게 블랙리스트에 등록 */
+        if (jwtTokenProvider.validateToken(token)) {
+            Date expirationDate = jwtTokenProvider.getExpirationDate(token, JwtTokenProvider.TokenType.ACCESS_TOKEN);
+            redisTemplate.opsForValue().set(
+                    token, true,
+                    expirationDate.getTime() - System.currentTimeMillis(), TimeUnit.MILLISECONDS); // 토큰의 유효기간이 지나면 자동 삭제
+            log.info("redis value : " + redisTemplate.opsForValue().get(token));
+        }
+
+        /* 새로운 access 토큰 발급 */
+        Authentication authentication = authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(user.getEmail(), newPassword)
+        );
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+        UserPrincipal newUserPrincipal = (UserPrincipal) authentication.getPrincipal();
+        String newAccessToken = jwtTokenProvider.generateAccessToken(newUserPrincipal);
+
+        return ResponseEntity.status(HttpStatus.OK).body(new AuthResponseDto(newAccessToken));
+    }
+
 
     /*@ApiOperation(value = "비밀번호 변경", notes = "회원의 비밀번호를 변경한다.", authorizations = { @Authorization(value = "jwtToken")})
     @ApiImplicitParam(name = "userPassword", value = "변경할 새로운 비밀번호", required = true, dataType = "string", defaultValue = "newpassword")
@@ -149,20 +258,4 @@ public class UserController {
     }*/
 
 
-//    // access 토큰이 만료되었으나 refresh 토큰이 살아있는 경우 재발급(그러나 필터통과를 못함(?)..)
-//    @ApiOperation(value = "토큰 만료", notes = "refresh 토큰을 이용하여 access 토큰을 재생성한다.", authorizations = { @Authorization(value = "jwtToken")})
-//    @PostMapping(value = "/check")
-//    public ResponseEntity<?> checkForGenerate(@AuthenticationPrincipal UserPrincipal userPrincipal,
-//                                              HttpServletRequest request) {
-//        String accessToken = jwtTokenProvider.extractToken(request);
-//
-//        if (!jwtTokenProvider.validateToken(accessToken)){
-//            String userEmail = userPrincipal.getEmail();
-//            if (redisTemplate.opsForValue().get(userEmail) != null) {
-//                String newAccessToken = jwtTokenProvider.generateAccessToken(userPrincipal);
-//                return ResponseEntity.ok(new AuthResponseDto(newAccessToken));
-//            }
-//        }
-//        return ResponseEntity.status(HttpStatus.OK).build();
-//    }
 }
