+++ src/main/java/com/automart/user/controller/UserSignController.java
@@ -1,13 +1,14 @@
 package com.automart.user.controller;
 
+
+import com.automart.user.dto.*;
+import com.automart.user.service.AdminService;
 import com.automart.advice.exception.*;
 import com.automart.security.UserPrincipal;
 import com.automart.security.jwt.JwtTokenProvider;
+import com.automart.user.domain.Admin;
 import com.automart.user.domain.AuthProvider;
 import com.automart.user.domain.User;
-import com.automart.user.dto.AuthResponseDto;
-import com.automart.user.dto.UserSignInRequestDto;
-import com.automart.user.dto.UserSignUpRequestDto;
 import com.automart.user.service.UserService;
 import io.swagger.annotations.*;
 import lombok.extern.slf4j.Slf4j;
@@ -24,6 +25,7 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.util.Collections;
 import java.util.Date;
@@ -32,8 +34,10 @@ import java.util.concurrent.TimeUnit;
 @Api(tags = {"1. Managing User Authentication "})
 @RestController
 @Slf4j
-@RequestMapping("/users")
 public class UserSignController {
+
+    @Autowired
+    private AdminService adminService;
 
     @Autowired
     private UserService userService;
@@ -54,7 +58,7 @@ public class UserSignController {
             @ApiResponse(code = 200, message = "이메일이 중복되지 않습니다."),
             @ApiResponse(code = 406, message = "동일한 이메일의 회원이 이미 존재합니다.")
     })
-    @PostMapping("/valid/email")
+    @PostMapping("/users/valid/email")
     public ResponseEntity<Void> validEmail(@RequestBody String email) throws ForbiddenSignUpException {
         userService.checkDuplicateEmail(email);
         return ResponseEntity.status(HttpStatus.OK).build();
@@ -68,7 +72,7 @@ public class UserSignController {
             @ApiResponse(code = 406, message = "올바른 형식의 이메일 주소가 아닙니다. ('@' 미포함)"),
             @ApiResponse(code = 428, message = "비밀번호를 변경해야 합니다.", response = AuthResponseDto.class)
     })
-    @PostMapping("/signin")
+    @PostMapping("/users/signin")
     public ResponseEntity<AuthResponseDto> signIn(@ApiParam("로그인 정보") @Valid @RequestBody UserSignInRequestDto requestDto) throws NotFoundUserException, SignInTypeErrorException {
         User user = userService.checkLogIn(requestDto.getEmail(), requestDto.getPassword());
 
@@ -108,7 +112,7 @@ public class UserSignController {
             @ApiResponse(code = 500, message = "메세지 전송 실패"),
             @ApiResponse(code = 406, message = "동일한 휴대폰 번호의 회원이 이미 존재합니다.")
     })
-    @PostMapping("/valid/phone")
+    @PostMapping("/users/valid/phone")
     public ResponseEntity<Integer> validPhoneForSignUp(@RequestParam String phoneNo) throws SMSException, ForbiddenSignUpException {
         userService.checkDuplicateTel(phoneNo);
         int validNum = userService.validatePhone(phoneNo);
@@ -121,7 +125,7 @@ public class UserSignController {
             @ApiResponse(code = 201, message = "정상적으로 회원가입이 완료되었습니다."),
             @ApiResponse(code = 406, message = "이메일 또는 핸드폰번호가 중복되어 회원가입에 실패하였습니다.")
     })
-    @PostMapping("/signup")
+    @PostMapping("/users/signup")
     public ResponseEntity<Void> signUp(@ApiParam("가입 회원 정보") @Valid @RequestBody UserSignUpRequestDto requestDto) throws ForbiddenSignUpException{
         userService.checkDuplicateEmail(requestDto.getEmail());
         userService.checkDuplicateTel(requestDto.getTel());
@@ -140,85 +144,58 @@ public class UserSignController {
     }
 
 
-    @ApiOperation(value = "8 로그인 후 비밀번호 변경", notes = "로그인 후 임시비밀번호를 변경한다.", authorizations = {@Authorization(value = "jwtToken")})
-    @ApiImplicitParam(name = "newPassword", value = "해당 회원의 새 비밀번호", required = true, dataType = "string", defaultValue = "newpassword")
+    @ApiOperation(value = "(개발용) 관리자 회원가입", notes = "개발자가 관리자를 추가하는데 사용, Front와 연결 X")
     @ApiResponses({
-            @ApiResponse(code = 201, message = "정상적으로 비밀번호를 변경 후, 새로운 토큰을 발급한다."),
-            @ApiResponse(code = 401, message = "토큰 만료로 인해 비밀번호 변경 불가 -> 새로운 토큰 발급", response = AuthResponseDto.class),
-            @ApiResponse(code = 403, message = "일치하는 회원이 존재하지 않아 비밀번호 변경에 실패하였습니다.")
+            @ApiResponse(code = 201, message = "정상적으로 회원가입이 완료되었습니다."),
+            @ApiResponse(code = 406, message = "아이디가 중복되어 회원가입에 실패하였습니다.")
     })
-    @PreAuthorize("hasRole('ROLE_USER')")
-    @PostMapping(value = "/change/password")
-    public ResponseEntity<AuthResponseDto> changePassword(@ApiIgnore @RequestHeader("Authorization") String token,
-                                                          @AuthenticationPrincipal UserPrincipal userPrincipal,
-                                                          @RequestParam String newPassword) throws NotFoundUserException {
-        /* password 변경 */
-        User user = userService.changePassword(userPrincipal.getNo(), newPassword);
-        user.makeFalseTempPw();
+    @PostMapping("/admin/signup")
+    public ResponseEntity<Void> adminSignUp(@ApiParam("가입 회원 정보") @Valid @RequestBody AdminSignUpRequestDto requestDto) throws ForbiddenSignUpException {
+        adminService.checkDuplicateId(requestDto.getId());
 
-        /* access token이 유효한 토큰인 경우 더 이상 사용하지 못하게 블랙리스트에 등록 */
-        if (jwtTokenProvider.validateToken(token)) {
-            Date expirationDate = jwtTokenProvider.getExpirationDate(token, JwtTokenProvider.TokenType.ACCESS_TOKEN);
-            redisTemplate.opsForValue().set(
-                    token, true,
-                    expirationDate.getTime() - System.currentTimeMillis(), TimeUnit.MILLISECONDS); // 토큰의 유효기간이 지나면 자동 삭제
-            log.info("redis value : " + redisTemplate.opsForValue().get(token));
-        }
+        Admin admin = Admin.builder()
+                .id(requestDto.getId())
+                .password(requestDto.getPassword())
+                .roles(Collections.singletonList("ROLE_ADMIN"))
+                .build();
 
-        /* 새로운 access 토큰 발급 */
+        adminService.saveAdmin(admin);
+
+        return ResponseEntity.status(HttpStatus.CREATED).build();
+    }
+
+
+    @ApiOperation(value = "1 관리자 로그인", notes = "관리자 로그인을 시도한다")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "정상적으로 로그인 되었습니다.", response = AuthResponseDto.class),
+            @ApiResponse(code = 403, message = "아이디 또는 비밀번호가 일치하지 않습니다."),
+            @ApiResponse(code = 406, message = "아이디에는 '@'가 포함될 수 없습니다.")
+    })
+    @PostMapping("/admin/signin")
+    public ResponseEntity<AuthResponseDto> adminSignIn(@ApiParam("로그인 정보") @Valid @RequestBody AdminSignInRequestDto requestDto) throws NotFoundUserException, SignInTypeErrorException {
+        /* '@'포함여부, 관리자 일치여부 체크 */
+        Admin admin = adminService.checkLogIn(requestDto.getId(), requestDto.getPassword());
+
         Authentication authentication = authenticationManager.authenticate(
-                new UsernamePasswordAuthenticationToken(user.getEmail(), newPassword)
+                new UsernamePasswordAuthenticationToken(
+                        requestDto.getId(),
+                        requestDto.getPassword()
+                )
         );
         SecurityContextHolder.getContext().setAuthentication(authentication);
-        UserPrincipal newUserPrincipal = (UserPrincipal) authentication.getPrincipal();
-        String newAccessToken = jwtTokenProvider.generateAccessToken(newUserPrincipal);
+        UserPrincipal userPrincipal = (UserPrincipal) authentication.getPrincipal();
 
-        return ResponseEntity.status(HttpStatus.OK).body(new AuthResponseDto(newAccessToken));
+        /* access 토큰과 refresh 토큰을 발급 */
+        String accessToken = jwtTokenProvider.generateAccessToken(userPrincipal);
+        String refreshToken = jwtTokenProvider.generateRefreshToken(userPrincipal);
+
+        /* refresh 토큰을 redis에 저장 */
+        Date expirationDate = jwtTokenProvider.getExpirationDate(refreshToken, JwtTokenProvider.TokenType.REFRESH_TOKEN);
+        redisTemplate.opsForValue().set(
+                userPrincipal.getPrincipal(), refreshToken,
+                expirationDate.getTime() - System.currentTimeMillis(), TimeUnit.MILLISECONDS); // 토큰의 유효기간이 지나면 자동 삭제
+        log.info("redis value : " + redisTemplate.opsForValue().get(userPrincipal.getPrincipal()));
+
+        return ResponseEntity.ok(new AuthResponseDto(accessToken));
     }
-
-    @ApiOperation(value = "(개발용) 회원 탈퇴", notes = "특정 회원을 삭제한다", authorizations = { @Authorization(value = "jwtToken")})
-    @ApiResponses({
-            @ApiResponse(code = 200, message = "회원이 정상적으로 탈퇴되었습니다.")
-    })
-    @DeleteMapping(value = "/delete")
-    public ResponseEntity<String> withdraw(@ApiIgnore @RequestHeader("Authorization") String token) {
-        return ResponseEntity.status(HttpStatus.OK).body(userService.withdraw(token, JwtTokenProvider.TokenType.ACCESS_TOKEN));
-    }
-
-
-    /*@ApiOperation(value = "로그아웃", notes = "로그인된 계정을 로그아웃한다.", authorizations = { @Authorization(value = "jwtToken")})
-    @ApiResponse(code = 200, message = "로그아웃 되었습니다.")
-    //@PreAuthorize("hasRole('USER')")
-    @PostMapping(value = "/logout")
-    public ResponseEntity<Void> logout(@AuthenticationPrincipal UserPrincipal userPrincipal,
-                                       HttpServletRequest request) {
-        String accessToken = jwtTokenProvider.extractToken(request);
-        String userEmail = null;
-
-        *//* access token을 통해 userEmail을 찾아 redis에 저장된 refresh token을 삭제한다.*//*
-        try {
-            userEmail = userPrincipal.getEmail();
-        } catch (InvalidTokenException e) {
-            log.error("userEmail이 유효한 토큰에 존재하지 않음.");
-        }
-
-        try {
-            if (redisTemplate.opsForValue().get(userEmail) != null) {
-                redisTemplate.delete(userEmail);
-            }
-        } catch (IllegalArgumentException e) {
-            redisTemplate.delete(userEmail);
-        }
-
-        *//* access token이 유효한 토큰인 경우 더 이상 사용하지 못하게 블랙리스트에 등록 *//*
-        if (jwtTokenProvider.validateToken(accessToken)) {
-            Date expirationDate = jwtTokenProvider.getExpirationDate(accessToken, JwtTokenProvider.TokenType.ACCESS_TOKEN);
-            redisTemplate.opsForValue().set(
-                    accessToken, true,
-                    expirationDate.getTime() - System.currentTimeMillis(), TimeUnit.MILLISECONDS); // 토큰의 유효기간이 지나면 자동 삭제
-            log.info("redis value : " + redisTemplate.opsForValue().get(accessToken));
-        }
-
-        return ResponseEntity.status(HttpStatus.OK).build();
-    }*/
 }
