+++ src/main/java/com/automart/user/dto/AdminSignInRequestDto.java
@@ -1,10 +1,11 @@
-package com.automart.admin.dto;
+package com.automart.user.dto;
 
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 public class AdminSignInRequestDto {
+
     private String id; // 관리자 아이디
     private String password; // 관리자 패스워드
